Reorganize sign/verify as panels in main window
